Remove function to exit when even queue is empty.
The function could easily create a deadlock. If called from the main
thread it will block until the event loop is empty, which will never
happen because the event loop is blocked.

Instead, just call `std::process::exit` and don't rely on functions
posted to the event loop to complete.