Remove window lookup on each call to `WindowHandle` functions.
Previously, a WindowHandle kept a ContextHandle and a WindowId,
and every time the window had to be queried or modified, the window
would be looked up again by it's ID.

Now, the lookup is done once and the WindowId is replaced by an index.
This also means that most functions no longer return a `Result`,
because the WindowHandle is guaranteed to be valid.

However, in order to keep that guarantee, we can no longer create or
delete windows as long as a WindowHandle exists.
That also means you can not get a `&mut ContextHandle` from a
`&mut WindowHandle` anymore.